+++ example/augmentation/main.go
@@ -2,13 +2,80 @@ package main
 
 import (
 	"fmt"
+	"log"
 
 	"github.com/sugarme/gotch"
 	"github.com/sugarme/gotch/vision"
 	"github.com/sugarme/gotch/vision/aug"
+	// ts "github.com/sugarme/gotch/tensor"
 )
 
 func main() {
+
+	// roundTrip()
+	tOne()
+
+}
+
+func roundTrip() {
+	img, err := vision.Load("./bb.png")
+	if err != nil {
+		panic(err)
+	}
+
+	fmt.Printf("%i", img)
+	fimg := aug.Byte2FloatImage(img)
+	fmt.Printf("%i", fimg)
+
+	bimg := aug.Float2ByteImage(fimg)
+	fmt.Printf("%i", bimg)
+
+	err = vision.Save(bimg, "./bimg.png")
+	if err != nil {
+		log.Fatal(err)
+	}
+}
+
+func tOne() {
+	img, err := vision.Load("./bb.png")
+	if err != nil {
+		panic(err)
+	}
+
+	// device := gotch.CudaIfAvailable()
+	device := gotch.CPU
+	imgTs := img.MustTo(device, true)
+
+	t, err := aug.Compose(aug.WithRandomSolarize(aug.WithSolarizeThreshold(125), aug.WithSolarizePvalue(1.0)))
+	// t, err := aug.Compose(aug.WithRandomAdjustSharpness(aug.WithSharpnessPvalue(1.0), aug.WithSharpnessFactor(10)))
+	// t, err := aug.Compose(aug.WithRandRotate(0, 360))
+	// t, err := aug.Compose(aug.WithResize(320, 320)) // NOTE. WithResize just works on CPU.
+	// t, err := aug.Compose(aug.WithRandomPosterize(aug.WithPosterizeBits(2), aug.WithPosterizePvalue(1.0)))
+	// t, err := aug.Compose(aug.WithRandomPerspective(aug.WithPerspectiveScale(0.6), aug.WithPerspectivePvalue(1.0)))
+	// t, err := aug.Compose(aug.WithNormalize(aug.WithNormalizeMean([]float64{0.485, 0.456, 0.406}), aug.WithNormalizeStd([]float64{0.229, 0.224, 0.225})))
+	// t, err := aug.Compose(aug.WithRandomInvert(1.0))
+	// t, err := aug.Compose(aug.WithRandomGrayscale(1.0))
+	// t, err := aug.Compose(aug.WithRandomVFlip(1.0))
+	// t, err := aug.Compose(aug.WithRandomHFlip(1.0))
+	// t, err := aug.Compose(aug.WithRandomEqualize(1.0))
+	// t, err := aug.Compose(aug.WithRandomCutout(aug.WithCutoutValue([]int64{124, 96, 255}), aug.WithCutoutScale([]float64{0.01, 0.1}), aug.WithCutoutRatio([]float64{0.5, 0.5})))
+	// t, err := aug.Compose(aug.WithCenterCrop([]int64{320, 320}))
+	// t, err := aug.Compose(aug.WithRandomAutocontrast())
+	// t, err := aug.Compose(aug.WithColorJitter(0.3, 0.3, 0.3, 0.3))
+	// t, err := aug.Compose(aug.WithGaussianBlur([]int64{5, 5}, []float64{1.0, 2.0}))
+	// t, err := aug.Compose(aug.WithRandomAffine(aug.WithAffineDegree([]int64{0, 15}), aug.WithAffineShear([]float64{0, 15})))
+
+	out := t.Transform(imgTs)
+	fname := fmt.Sprintf("./bb-transformed.png")
+	err = vision.Save(out, fname)
+	if err != nil {
+		panic(err)
+	}
+	imgTs.MustDrop()
+	out.MustDrop()
+}
+
+func tMany() {
 	n := 360
 	for i := 1; i <= n; i++ {
 		img, err := vision.Load("./bb.png")
@@ -19,22 +86,6 @@ func main() {
 		// device := gotch.CudaIfAvailable()
 		device := gotch.CPU
 		imgTs := img.MustTo(device, true)
-		// t, err := aug.Compose(aug.WithResize(512, 512)) // NOTE. WithResize just works on CPU.
-		// t, err := aug.Compose(aug.WithRandRotate(0, 360), aug.WithColorJitter(0.3, 0.3, 0.3, 0.4))
-		// t, err := aug.Compose(aug.WithGaussianBlur([]int64{5, 5}, []float64{1.0, 2.0}), aug.WithRandRotate(0, 360), aug.WithColorJitter(0.3, 0.3, 0.3, 0.3))
-		// t, err := aug.Compose(aug.WithRandomCrop([]int64{320, 320}, []int64{10, 10}, true, "constant"))
-		// t, err := aug.Compose(aug.WithCenterCrop([]int64{320, 320}))
-		// t, err := aug.Compose(aug.WithRandomCutout(aug.WithCutoutValue([]int64{124, 96, 255}), aug.WithCutoutScale([]float64{0.01, 0.1}), aug.WithCutoutRatio([]float64{0.5, 0.5})))
-		// t, err := aug.Compose(aug.WithRandomPerspective(aug.WithPerspectiveScale(0.6), aug.WithPerspectivePvalue(0.8)))
-		// t, err := aug.Compose(aug.WithRandomAffine(aug.WithAffineDegree([]int64{0, 15}), aug.WithAffineShear([]float64{0, 15})))
-		// t, err := aug.Compose(aug.WithRandomGrayscale(0.5))
-		// t, err := aug.Compose(aug.WithRandomSolarize(aug.WithSolarizeThreshold(125), aug.WithSolarizePvalue(0.5)))
-		// t, err := aug.Compose(aug.WithRandomInvert(0.5))
-		// t, err := aug.Compose(aug.WithRandomPosterize(aug.WithPosterizeBits(2), aug.WithPosterizePvalue(1.0)))
-		// t, err := aug.Compose(aug.WithRandomAutocontrast())
-		// t, err := aug.Compose(aug.WithRandomAdjustSharpness(aug.WithSharpnessPvalue(0.3), aug.WithSharpnessFactor(10)))
-		// t, err := aug.Compose(aug.WithRandomEqualize(1.0))
-		// t, err := aug.Compose(aug.WithNormalize(aug.WithNormalizeMean([]float64{0.485, 0.456, 0.406}), aug.WithNormalizeStd([]float64{0.229, 0.224, 0.225})))
 
 		t, err := aug.Compose(
 			aug.WithResize(200, 200),
@@ -67,4 +118,5 @@ func main() {
 
 		fmt.Printf("%03d/%v completed.\n", i, n)
 	}
+
 }